requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
python-jose[cryptography]
bcrypt
